Adds code for custom parent slug field
This is necessary because entity tokens (the natural/easy solution to
building a slug based on some attribute of a related entity) don't work
properly on create of a node, meaning that when a collection is saved
its slug is broken/wrong until that collection is saved again.  Since
this is a pretty elementary fail handling it with some hackiness seemed
preferrable to ignoring it.